refactor(dashboard): extract memory_metrics + ooda_thinking to metrics.rs (#1266)
routes.rs: 2517 -> 2378
metrics.rs: NEW 145

Promotes file_metrics + count_json_records to pub(crate) so
metrics module can call them.
